Added notifications system to follow logic

+++ server/src/Application/Follows/Commands/UnfollowUser/UnfollowUser.cs
@@ -1,5 +1,6 @@
 ﻿using Application.Common.Abstractions;
 using Application.Common.Interfaces;
+using Domain.Exceptions;
 
 namespace Application.Follows.Commands.UnfollowUser;
 
@@ -29,8 +30,26 @@ public class UnfollowUserCommandHandler : ICommandHandler<UnfollowUserCommand, U
         {
             return new UnfollowUserResponse(false);
         }
-
+        
+        var currentUser = await _context.Users.FindAsync(new object[] { currentUserId }, cancellationToken);
+        var user = await _context.Users.FindAsync(new object[] { request.UserId }, cancellationToken);
+        
+        if (currentUser is null || user is null)
+        {
+            throw new NotFoundException("User not found");
+        }
+        
+        currentUser.FollowingCount--;
+        user.FollowersCount--;
         _context.Follows.Remove(follow);
+        
+        var existingNotification = await _context.Notifications
+            .FirstOrDefaultAsync(n => n.UserId == request.UserId && n.SenderId == currentUserId, cancellationToken);
+        
+        if (existingNotification is not null)
+        {
+            _context.Notifications.Remove(existingNotification);
+        }
 
         await _context.SaveChangesAsync(cancellationToken);
